requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.3
langchain>=0.0.267
langchain-mistralai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
plotly>=5.15.0
